feature: support parse OpenRTK and RTK330LA log files
