+++ src/aceinna/bootstrap/loader.py
@@ -5,6 +5,7 @@ Application Loader
 from .receiver import Receiver as ReceiverApp
 from .cli import CommandLine as CommandLineApp
 from .default import Default as DefaultApp
+from .log_parser import LogParser as LogParserApp
 from .. import VERSION
 
 
@@ -12,7 +13,7 @@ class APP_TYPE:
     DEFAULT = 'default'
     CLI = 'cli'
     RECEIVER = 'receiver'
-    PARSER = 'parser'
+    LOG_PARSER = 'log-parser'
 
 
 class Loader:
@@ -33,6 +34,9 @@ class Loader:
 
         if platform == APP_TYPE.RECEIVER:
             active_app = ReceiverApp(**options)
+
+        if platform == APP_TYPE.LOG_PARSER:
+            active_app = LogParserApp(**options)
 
         if active_app is None:
             raise ValueError('no matched bootstrap')
