+++ src/aceinna/bootstrap/log_parser.py
@@ -1,0 +1,43 @@
+import os
+from ..tools.openrtk_parse import do_parse as do_parse_openrtk_logs
+from ..tools.rtkl_parse import do_parse as do_parse_rtkl_logs
+from ..models import LogParserArgs
+
+
+class LogParser:
+    _options = None
+    _tunnel = None
+    _driver = None
+
+    def __init__(self, **kwargs):
+        self._build_options(**kwargs)
+
+    def listen(self):
+        '''Start to do parse
+        '''
+        self._validate_params()
+
+        setting_file = 'log-parser.json'
+
+        if self._options.log_type == 'openrtk':
+            do_parse_openrtk_logs(self._options.path,
+                                  self._options.kml_rate,
+                                  setting_file)
+        elif self._options.log_type == 'rtkl':
+            do_parse_rtkl_logs(self._options.path,
+                               self._options.kml_rate,
+                               setting_file)
+        else:
+            raise ValueError('No matched log parser')
+
+        os._exit(1)
+
+    def _validate_params(self):
+        for attr_name in ['log_type', 'path', 'kml_rate']:
+            attr_value = getattr(self._options, attr_name)
+            if not attr_value:
+                raise ValueError(
+                    'Parameter {0} should have a value'.format(attr_name))
+
+    def _build_options(self, **options):
+        self._options = LogParserArgs(**options)
