+++ src/aceinna/models/__init__.py
@@ -1,2 +1,3 @@
 from .args import WebserverArgs
+from .args import LogParserArgs
 from .internal_combine_app_parse_rule import InternalCombineAppParseRule
